Move functions from `util.js` to specialized files
