+++ src/routes/detail-page.js
@@ -1,14 +1,15 @@
 const Apify = require('apify');
+const { REVIEWS_RESULTS_PER_REQUEST, LABELS, PLACE_COUNTRY_URL_CODE_REGEX } = require('../consts');
 const { extractDetail, extractPreviewReviews } = require('../extraction/detail-page-extraction');
-const { getMaxReviewsPages, addDetail } = require('../global-store');
+const { getMaxReviewsPages, addDetail, setReviewUrlsToProcess } = require('../global-store');
 
 const {
     validateProxy,
     saveDetailIfComplete,
-    enqueueAllReviewsPages,
     isObject,
     getAttribute,
     getPagename,
+    getLocalizedUrl,
 } = require('../util');
 
 const { log } = Apify.utils;
@@ -88,4 +89,81 @@ const getExtendedUserResult = async (page, extendOutputFunction, stringifiedExte
     }
 
     return userResult;
+};
+
+const enqueueAllReviewsPages = async (context, detailPagename, reviewsCount, language) => {
+    const { page, crawler: { requestQueue } } = context;
+
+    const detailPageUrl = await page.url();
+    const reviewsUrl = buildReviewsStartUrl(detailPageUrl, language);
+
+    const reviewPagesUrls = getReviewPagesUrls(reviewsUrl, reviewsCount);
+    log.info(`Found ${reviewsCount} reviews.
+    Enqueuing ${reviewPagesUrls.length} reviews pages (${REVIEWS_RESULTS_PER_REQUEST} reviews per page)...`, { detailPageUrl });
+
+    setReviewUrlsToProcess(detailPagename, reviewPagesUrls);
+
+    for (let index = 0; index < reviewPagesUrls.length; index++) {
+        const url = reviewPagesUrls[index];
+        const request = {
+            url,
+            userData: {
+                label: LABELS.REVIEW,
+                detailPagename,
+            },
+        };
+
+        await requestQueue.addRequest(
+            request,
+
+            /**
+            * Reviews have to be prioritized as we're waiting for all
+            * reviews to be processed before we push the detail into the dataset.
+            */
+            { forefront: true },
+        );
+    }
+};
+
+const buildReviewsStartUrl = (detailPageUrl, language) => {
+    const url = new URL(detailPageUrl);
+    const { searchParams } = url;
+
+    const reviewsBaseUrl = 'https://www.booking.com/reviewlist.html';
+    const reviewsUrl = new URL(getLocalizedUrl(reviewsBaseUrl, language));
+
+    // regex.exec(string) needs to be used instead of string.match(regex) to make capturing group work properly
+    const placeCountryMatches = PLACE_COUNTRY_URL_CODE_REGEX.exec(detailPageUrl);
+    const placeCountryMatch = placeCountryMatches ? placeCountryMatches[1] : '';
+
+    const reviewUrlParams = {
+        aid: searchParams.get('aid'),
+        label: searchParams.get('label'),
+        sid: searchParams.get('sid'),
+        srpvid: searchParams.get('srpvid'),
+        pagename: getPagename(detailPageUrl),
+        cc1: placeCountryMatch,
+        rows: REVIEWS_RESULTS_PER_REQUEST,
+        offset: 0,
+    };
+
+    Object.keys(reviewUrlParams).forEach((key) => {
+        reviewsUrl.searchParams.set(key, reviewUrlParams[key]);
+    });
+
+    return reviewsUrl;
+};
+
+const getReviewPagesUrls = (reviewsUrl, reviewsCount) => {
+    const urlsToEnqueue = [];
+
+    const maxReviewsPages = getMaxReviewsPages();
+    const reviewsToEnqueue = Math.min(reviewsCount, maxReviewsPages * REVIEWS_RESULTS_PER_REQUEST);
+
+    for (let enqueuedReviews = 0; enqueuedReviews < reviewsToEnqueue; enqueuedReviews += REVIEWS_RESULTS_PER_REQUEST) {
+        reviewsUrl.searchParams.set('offset', enqueuedReviews);
+        urlsToEnqueue.push(reviewsUrl.toString());
+    }
+
+    return urlsToEnqueue;
 };
