+++ src/util.js
@@ -6,22 +6,9 @@ const { GlobalStore } = require('apify-global-store');
 const {
     DATE_FORMAT,
     PROPERTY_TYPE_IDS,
-    RESULTS_PER_PAGE,
     PLACE_URL_NAME_REGEX,
-    REVIEWS_RESULTS_PER_REQUEST,
-    PLACE_COUNTRY_URL_CODE_REGEX,
-    LABELS,
     LOCALIZATION_REGEX,
 } = require('./consts');
-
-const {
-    addEnqueuedUrl,
-    decrementRemainingPages,
-    setReviewUrlsToProcess,
-    getRemainingPages,
-    getEnqueuedUrls,
-    getMaxReviewsPages,
-} = require('./global-store');
 
 const { log } = Apify.utils;
 
@@ -280,166 +267,7 @@ module.exports.checkDateGap = (checkIn, checkOut) => {
     return -1;
 };
 
-/**
- * Extracts information from the detail page and enqueue all pagination pages.
- *
- * @param {Page} page - The Puppeteer page object.
- * @param {RequestQueue} requestQueue - RequestQueue to add the requests to.
- * @param {{ input: Object }} globalContext - Actor's global context.
- */
-module.exports.enqueueAllPaginationPages = async (page, requestQueue, globalContext) => {
-    const { input } = globalContext;
-
-    const baseUrl = page.url();
-    if (baseUrl.indexOf('offset') < 0) {
-        log.info('enqueuing pagination pages...');
-        const countSelector = '.sorth1, .sr_header h1, .sr_header h2, [data-capla-component*="HeaderDesktop"] h1';
-        try {
-            const pageUrl = await page.url();
-            await page.waitForSelector(countSelector);
-            const countElem = await page.$(countSelector);
-            const countData = (await getAttribute(countElem, 'textContent')).replace(/\.|,|\s/g, '').match(/\d+/);
-
-            if (countData) {
-                const count = Math.ceil(parseInt(countData[0], 10) / RESULTS_PER_PAGE);
-                log.info(`pagination pages: ${count}`);
-
-                for (let i = 1; i < count; i++) {
-                    const newOffset = RESULTS_PER_PAGE * i;
-                    const newUrl = pageUrl.includes('offset=')
-                        ? pageUrl.replace(/offset=(\d+)/, `offset=${newOffset}`)
-                        : `${pageUrl}&offset=${newOffset}`;
-
-                    if (getRemainingPages() < 1) {
-                        break;
-                    }
-
-                    decrementRemainingPages();
-
-                    await requestQueue.addRequest({
-                        url: addUrlParameters(newUrl, input),
-                        userData: { label: 'page' },
-                    });
-                }
-            }
-        } catch (e) {
-            log.warning(e);
-        }
-    }
-};
-
-module.exports.enqueueAllReviewsPages = async (context, detailPagename, reviewsCount, language) => {
-    const { page, crawler: { requestQueue } } = context;
-
-    const detailPageUrl = await page.url();
-    const reviewsUrl = buildReviewsStartUrl(detailPageUrl, language);
-
-    const reviewPagesUrls = getReviewPagesUrls(reviewsUrl, reviewsCount);
-    log.info(`Found ${reviewsCount} reviews.
-    Enqueuing ${reviewPagesUrls.length} reviews pages (${REVIEWS_RESULTS_PER_REQUEST} reviews per page)...`, { detailPageUrl });
-
-    setReviewUrlsToProcess(detailPagename, reviewPagesUrls);
-
-    for (let index = 0; index < reviewPagesUrls.length; index++) {
-        const url = reviewPagesUrls[index];
-        const request = {
-            url,
-            userData: {
-                label: LABELS.REVIEW,
-                detailPagename,
-            },
-        };
-
-        await requestQueue.addRequest(
-            request,
-
-            /**
-            * Reviews have to be prioritized as we're waiting for all
-            * reviews to be processed before we push the detail into the dataset.
-            */
-            { forefront: true },
-        );
-    }
-};
-
-module.exports.enqueueFilterLinks = async (extractionInfo, urlInfo, requestQueue) => {
-    const { page, unchecked, attribute } = extractionInfo;
-    const { label, baseUrl } = urlInfo;
-
-    const url = new URL(baseUrl);
-
-    const uncheckedElements = await page.$$(unchecked);
-    const uncheckedFilters = await getFilterNameValues(uncheckedElements, attribute);
-
-    const validFilters = getValidFilters(uncheckedFilters);
-    const filtersToEnqueue = getFiltersToEnqueue(validFilters, url);
-    const newFiltersCount = filtersToEnqueue.length;
-
-    if (newFiltersCount) {
-        log.info(`enqueuing pages with ${filtersToEnqueue.length} new filters set...`);
-    }
-
-    await enqueueFilters(filtersToEnqueue, requestQueue, label, baseUrl);
-};
-
 module.exports.isObject = (val) => typeof val === 'object' && val !== null && !Array.isArray(val);
-
-const buildReviewsStartUrl = (detailPageUrl, language) => {
-    const url = new URL(detailPageUrl);
-    const { searchParams } = url;
-
-    const reviewsBaseUrl = 'https://www.booking.com/reviewlist.html';
-    const reviewsUrl = new URL(getLocalizedUrl(reviewsBaseUrl, language));
-
-    // regex.exec(string) needs to be used instead of string.match(regex) to make capturing group work properly
-    const placeCountryMatches = PLACE_COUNTRY_URL_CODE_REGEX.exec(detailPageUrl);
-    const placeCountryMatch = placeCountryMatches ? placeCountryMatches[1] : '';
-
-    const reviewUrlParams = {
-        aid: searchParams.get('aid'),
-        label: searchParams.get('label'),
-        sid: searchParams.get('sid'),
-        srpvid: searchParams.get('srpvid'),
-        pagename: getPagename(detailPageUrl),
-        cc1: placeCountryMatch,
-        rows: REVIEWS_RESULTS_PER_REQUEST,
-        offset: 0,
-    };
-
-    Object.keys(reviewUrlParams).forEach((key) => {
-        reviewsUrl.searchParams.set(key, reviewUrlParams[key]);
-    });
-
-    return reviewsUrl;
-};
-
-const getReviewPagesUrls = (reviewsUrl, reviewsCount) => {
-    const urlsToEnqueue = [];
-
-    const maxReviewsPages = getMaxReviewsPages();
-    const reviewsToEnqueue = Math.min(reviewsCount, maxReviewsPages * REVIEWS_RESULTS_PER_REQUEST);
-
-    for (let enqueuedReviews = 0; enqueuedReviews < reviewsToEnqueue; enqueuedReviews += REVIEWS_RESULTS_PER_REQUEST) {
-        reviewsUrl.searchParams.set('offset', enqueuedReviews);
-        urlsToEnqueue.push(reviewsUrl.toString());
-    }
-
-    return urlsToEnqueue;
-};
-
-const getFilterNameValues = async (elements, attribute) => {
-    const nameValues = {};
-
-    for (const element of elements) {
-        const filterValue = await getAttribute(element, attribute);
-
-        const [name, value] = filterValue.split('=');
-        nameValues[name] = nameValues[name] || [];
-        nameValues[name].push(value);
-    }
-
-    return nameValues;
-};
 
 /**
  *
@@ -448,103 +276,6 @@ const getFilterNameValues = async (elements, attribute) => {
  */
 const convertMomentToDate = (time) => {
     return time.toDate().setHours(0, 0, 0, 0);
-};
-
-const getValidFilters = (uncheckedFilters) => {
-    const validFilters = { ...uncheckedFilters };
-
-    /* Exclude review_score from filter enqueuing. It has a strict value for each run that can not be changed. */
-    delete validFilters.review_score;
-
-    // Invalid filter that is scraped among other checkboxes.
-    delete validFilters['1'];
-
-    return validFilters;
-};
-
-const getFiltersToEnqueue = (filters, url) => {
-    const filtersToEnqueue = [];
-
-    Object.keys(filters).forEach((name) => {
-        const values = filters[name].filter((value) => {
-            // remove value which is included with the same parameter in current url
-            return value && !url.search.includes(`${name}=${value}`);
-        });
-
-        const isFilterEnqueued = isFilterAlreadyEnqueued(name, url);
-
-        if (!isFilterEnqueued) {
-            filtersToEnqueue.push({ name, values });
-        }
-    });
-
-    return filtersToEnqueue;
-};
-
-const isFilterAlreadyEnqueued = (name, url) => {
-    const updatedUrl = new URL(url);
-    updatedUrl.searchParams.set(name, 'example');
-
-    const enqueuedUrls = getEnqueuedUrls();
-
-    for (const enqueuedUrl of enqueuedUrls) {
-        if (haveSameQueryParamNames(updatedUrl, enqueuedUrl)) {
-            return true;
-        }
-    }
-
-    return false;
-};
-
-const haveSameQueryParamNames = (firstUrl, secondUrl) => {
-    if (firstUrl.pathname !== secondUrl.pathname) {
-        return false;
-    }
-
-    /* Cross validation with matching firstUrl -> secondUrl parameters
-       as well as secondUrl -> firstUrl parameters. The length of searchParams
-       iterable can not be checked directly without looping through it.
-    */
-
-    for (const key of firstUrl.searchParams.keys()) {
-        if (!secondUrl.searchParams.has(key)) {
-            return false;
-        }
-    }
-
-    for (const key of secondUrl.searchParams.keys()) {
-        if (!firstUrl.searchParams.has(key)) {
-            return false;
-        }
-    }
-
-    return true;
-};
-
-const enqueueFilters = async (filters, requestQueue, label, baseUrl) => {
-    for (const filter of filters) {
-        const { name, values } = filter;
-
-        // Enqueue filter with all possible values.
-        for (const value of values) {
-            const url = new URL(baseUrl);
-            url.searchParams.set(name, value);
-
-            // Check that url with the exact same filters and values doesn't exist already.
-            const enqueuedUrls = getEnqueuedUrls();
-            const remainingPages = getRemainingPages();
-
-            if (!enqueuedUrls.includes(url) && remainingPages > 0) {
-                addEnqueuedUrl(url);
-                decrementRemainingPages();
-
-                await requestQueue.addRequest({
-                    url: url.toString(),
-                    userData: { label },
-                });
-            }
-        }
-    }
 };
 
 /**
